Clean trace package from SQL-related code; Fix tests
