+++ trace/sql/sql_test.go
@@ -1,8 +1,13 @@
 package sql
 
 import (
-	"github.com/stretchr/testify/assert"
+	"context"
 	"testing"
+
+	"github.com/beatlabs/patron/trace"
+	opentracing "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/mocktracer"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestParseDSN(t *testing.T) {
@@ -32,4 +37,29 @@ func TestParseDSN(t *testing.T) {
 			assert.Equal(t, got, tc.want)
 		})
 	}
+}
+
+func TestSQLStartFinishSpan(t *testing.T) {
+	mtr := mocktracer.New()
+	opentracing.SetGlobalTracer(mtr)
+	c := connInfo{"instance", "name"}
+	tag := opentracing.Tag{Key: "key", Value: "value"}
+	sp, req := c.startSpan(context.Background(), "sa", "ssf", tag)
+	assert.NotNil(t, sp)
+	assert.NotNil(t, req)
+	assert.IsType(t, &mocktracer.MockSpan{}, sp)
+	jsp := sp.(*mocktracer.MockSpan)
+	assert.NotNil(t, jsp)
+	trace.SpanSuccess(sp)
+	rawSpan := mtr.FinishedSpans()[0]
+	assert.Equal(t, map[string]interface{}{
+		"component":    "sql",
+		"version":      "dev",
+		"db.instance":  "instance",
+		"db.statement": "ssf",
+		"db.type":      "RDBMS",
+		"db.user":      "name",
+		"error":        false,
+		"key":          "value",
+	}, rawSpan.Tags())
 }
