+++ trace/trace_test.go
@@ -16,7 +16,7 @@ func TestSetup_Tracer_Close(t *testing.T) {
 	assert.NoError(t, err)
 	err = Close()
 	assert.NoError(t, err)
-	version = "dev"
+	Version = "dev"
 }
 
 func TestStartFinishConsumerSpan(t *testing.T) {
@@ -102,30 +102,6 @@ func TestHTTPStartFinishSpan(t *testing.T) {
 		"http.url":         "/",
 		"version":          "dev",
 		"correlationID":    "corID",
-	}, rawSpan.Tags())
-}
-
-func TestSQLStartFinishSpan(t *testing.T) {
-	mtr := mocktracer.New()
-	opentracing.SetGlobalTracer(mtr)
-	tag := opentracing.Tag{Key: "key", Value: "value"}
-	sp, req := SQLSpan(context.Background(), "name", "sql", "rdbms", "instance", "sa", "ssf", tag)
-	assert.NotNil(t, sp)
-	assert.NotNil(t, req)
-	assert.IsType(t, &mocktracer.MockSpan{}, sp)
-	jsp := sp.(*mocktracer.MockSpan)
-	assert.NotNil(t, jsp)
-	SpanSuccess(sp)
-	rawSpan := mtr.FinishedSpans()[0]
-	assert.Equal(t, map[string]interface{}{
-		"component":    "sql",
-		"version":      "dev",
-		"db.instance":  "instance",
-		"db.statement": "ssf",
-		"db.type":      "rdbms",
-		"db.user":      "sa",
-		"error":        false,
-		"key":          "value",
 	}, rawSpan.Tags())
 }
